+++ src/ITrackerContext.ts
@@ -1,0 +1,9 @@
+import { TrackedObject } from "./TrackedObject";
+
+export interface ITrackerContext {
+  isDirty: boolean;
+  isValid: boolean;
+  canCommit: boolean;
+  trackedObjects: TrackedObject[];
+  deletedObjects: TrackedObject[];
+}